CRUD/requirements.txt
autopep8
flask
mysql-connector-python